Fix: 홈 헤더 및 login 처리 수정
