+++ src/app/[locale]/page.tsx
@@ -1,20 +1,23 @@
 import HomeFooterButton from '@src/components/screens/home/components/HomeFooterButton';
 import HomeBody from '@src/components/screens/home/section/HomeBody';
 import HomeFooter from '@src/components/screens/home/section/HomeFooter';
-import HomeHeader from '@src/components/screens/home/section/HomeHeader';
-import { getCookie } from 'cookies-next';
-import { cookies } from 'next/headers';
+import dynamic from 'next/dynamic';
 
-const HomePage = () => {
-    const token = getCookie('refreshToken', { cookies });
-    return (
-        <div className="w-screen">
-            <HomeHeader isLoginSns={!!token} />
-            <HomeBody />
-            <HomeFooter />
-            <HomeFooterButton />
-        </div>
-    );
-};
+const HomeHeaderButton = dynamic(
+    () => import('@src/components/screens/home/section/HomeHeaderButton'),
+    {
+        ssr: false,
+    },
+);
 
+const HomePage = () => (
+    <div className="w-screen">
+        <header className="m-5 flex h-10 flex-row-reverse">
+            <HomeHeaderButton />
+        </header>
+        <HomeBody />
+        <HomeFooter />
+        <HomeFooterButton />
+    </div>
+);
 export default HomePage;
