+++ src/components/screens/home/components/login/HomeBasicLoginSection.tsx
@@ -1,7 +1,6 @@
-import { FC, useState } from 'react';
+import { FC } from 'react';
 import { postLogin } from '@src/api/postLogin';
 import Button from '@src/components/common/Button';
-import CommonInput from '@src/components/common/CommonInput';
 import Modal from '@src/components/common/modal/Modal';
 import defaultRequest from '@src/lib/axios/defaultRequest';
 import { ModalService } from '@src/service/ModalService';
@@ -16,16 +15,12 @@ import HomeBasicLoginSubmitButton from './HomeBasicLoginSubmitButton';
 import HomePasswordFindScreen from '../password-find/HomePasswordFindScreen';
 import HomeSignupScreen from '../signup/HomeSignupScreen';
 
-interface HomeBasicLoginSectionProps {
-    onSuccessLogin: VoidFunction;
-}
-
 export interface HomeBasicLoginFormValue {
     email: string;
     password: string;
 }
 
-const HomeBasicLoginSection: FC<HomeBasicLoginSectionProps> = ({ onSuccessLogin }) => {
+const HomeBasicLoginSection: FC = () => {
     const modalService = ModalService.getInstance();
     const toastService = ToastService.getInstance();
     const login = useMutation({
@@ -36,7 +31,6 @@ const HomeBasicLoginSection: FC<HomeBasicLoginSectionProps> = ({ onSuccessLogin 
                 const refreshToken = data.headers.refresh_token;
                 defaultRequest.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
                 await setCookie('refreshToken', refreshToken);
-                await onSuccessLogin();
                 toastService.addToast('로그인 되었습니다.');
                 modalService.closeEntireModal();
             }
@@ -70,7 +64,7 @@ const HomeBasicLoginSection: FC<HomeBasicLoginSectionProps> = ({ onSuccessLogin 
     const onClickSignup = () => {
         modalService.openModal(
             <Modal onClose={onCloseModal}>
-                <HomeSignupScreen onSuccessLogin={onSuccessLogin} />
+                <HomeSignupScreen />
             </Modal>,
         );
     };
